add menu section to plugins props
